feat: Started 2D Renderer This commit contains the start of the 2D Renderer which essentially abstracts away all the openGL buffer things and provides really robust functions instead.
Due to this being a large undertaking it is currently unfinished and the code DOES NOT COMPILE.

I also plan to hopefully turn this into a batch rendering system so that there are less draw calls per frame. However, this might be impossible if the time limit for the game jam does not allow it.